pyproject.toml
[project]
name = 'swelancer'
version = "0.1.0"
dependencies = [
    "alcatraz",
    "nanoeval",
    "nanoeval-alcatraz",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.9.0",
]
requires-python = ">=3.11"

[tool.setuptools]
packages = ["swelancer"]

[tool.uv.sources]
nanoeval = { path = "project/nanoeval", editable = true }
alcatraz = { path = "project/alcatraz", editable = true }
nanoeval-alcatraz = { path = "project/nanoeval_alcatraz", editable = true }
